refactor: split workspace.go into cohesive package files (#95)
Keep Service methods, locking, rollback, and tests unchanged. Move create,
remove, rename, sync, status, and doctor onto package-local files so
safety-sensitive mutation paths have a smaller review surface.
